refactor(ui): split chat pane layout rendering
